編集で両方いじるとクラッシュするバグを修正
